[refactor] Refactor iliexporter and iliimport
